Remove some unneeded auxpow logic.
In Xaya, auxpow handling (as part of triple-purpose mining) is simpler
than in Namecoin.  It is always active, and the chain ID is only verified
as part of the Merkle tree position; there are no explicit checks that need
the chain ID to be present in the auxpow itself.

This commit removes some logic that is not needed for Xaya due to these
reasons, simplifying the code for further work on PoW data.