Straighten auxpow logic in verify_header.
This makes the control flow in Blockchain.verify_header more straight
forward for the auxpow handling.  Previously, it was split up into
three parts (first testing for checkpoints, then computing the PoW hash
and then, later in the function, checking the PoW hash against the
target).  Now, all the PoW code is at the end of the function, after
the code that is in the upstream function.

Also includes a fix for a typo in the unit tests, which was previously
hidden by the control flow but is now relevant.