Use Xaya's difficulty algorithm in blockchain.py.
This adapts the blockchain-handling code to use Xaya's continuous and
dual-algo difficulty retargeting.  It requires some non-trivial changes,
since the assumption of one difficulty for each chunk is no longer
true.  Also the minimum difficulty is lower in Xaya (due to Neoscrypt
being harder than SHA-256d).

With this, Electrum syncs up to the Xaya mainnet chain.  Chainwork is
broken for now, though, as are checkpoints (but we don't have any
for now anyway).