Automatically convert addresses in tests.
The unit tests contain lots of addresses and private keys as strings.
Those are encoded in the Bitcoin formats in the upstream code, but need
to be converted to Namecoin formats to be accepted by the code under test.

This used to be done by hand using e.g. the conversion tool from Namecoin
Core.  But that is cumbersome.  In this change, we introduce a Python
function "frombtc" which handles the conversion automatically.  So instead
of explicitly converting string literals, we now only have to wrap them
in "frombtc(...)" in the tests.  This makes them easier to understand, and
also makes upstream merges easier (although it does not completely avoid
merge conflicts).

In this commit, the change is just applied to test_bitcoin.py as a proof
of concept.  More will come later.