Basic update of name logic to Xaya.
This is the minimal required update of the name logic to Xaya.  Mainly
we have to remove name expiration, and switch over to name_register
from the name_new/name_firstupdate pair (also getting rid of the need
for autoregister with queued firstupdates).

This does not update the GUI completely, only the parts necessary to
make it run at all.  The commands and basic functionality works, though.