Fix signature creation for segwit/P2SH names.
When segwit outputs are spent, BIP143 comes into play to determine how
the signature is made.  In particular, the hashed data that is signed
for each input also contains the input value.  For name inputs, that
should be the locked amount.  Since Electrum sets the value to zero
when parsing an output, we have to explicitly set it back to 0.01 NMC
for the signature.

This also fixes another compatibility issue between signatures created
in Electrum and those expected in Namecoin Core:  The scriptCode part
of the hashed input data should not contain the name prefix for
"indirected" input types (like P2SH or segwit), while the name prefix
is contained for ordinary P2PK and P2PKH inputs.