Add compatibility JSON-RPC server.
This adds a second JSON-RPC server that can be run optionally by the
Electrum daemon (namely if compatibilityrpcport is set in the config).
This server implements a few of the RPC methods most relevant to games
in a way that matches Xaya Core, so that games can use Electrum as a
drop-in replacement instead.

The methods supported are getbalance, getnewaddress, name_list,
name_register, name_update and name_pending.  name_pending only returns
an empty set for now; a proper implementation (e.g. based on a remote
server or if possible querying the mempool on ElectrumX) will be done
at a later stage.

This implements https://github.com/xaya/electrum-chi/issues/4.