Address conversion for test_storage_upgrade.py.
Do automatic address conversion also in test_storage_upgrade.py.  Since
this contains lots of large strings that have many addresses in them,
we use a custom-made function for recursive conversion of JSON objects
instead of manually adding frombtc(...) as in the other tests.

Note that this has the effect that we now convert all addresses, while
previously only those tests were converted that actually failed when left
in the original version.