Update rebranding check for address conversion.
With the recent batch of address-conversion commits to the tests we broke
the rebranding check, because e.g. "bitcoin:" appears now in those tests
literally.  This updates the rebranding check to still succeed.

We also simplify a complex regexp that was (according to comments)
"designed" to exclude certain things by just explicitly excluding those.