Clean up deserialisation code for block headers.
This overhauls the code for deserialising block headers (including with
AuxPoW).  It removes the "expect_trailing_data" optional feature except
where it is really needed, and splits parsing of "pure" headers without
AuxPoW and "full" headers with AuxPoW into two separate functions as
suggested in https://github.com/namecoin/electrum-nmc/issues/154.

Besides making the code simpler, this also forces callers to explicitly
specify whether they expect a pure or full block header; that helps to
make the code easier to understand, and also to avoid unintended
behaviour.

This should also improve upstream merges for two reasons:  First, the new
deserialize_pure_header function is a literal copy of the upstream
deserialize_header, just renamed.  And second, due to the rename, it will
now be easy to detect new calls, so that we are forced to consciously
handle them and decide which of the two new functions to use.

Finally, the AuxPoW tests are extended with situations where we make use
of expect_trailing_data and start_position to ensure that those two work
as they should when used.