Add optional outputs argument to name commands.
This adds a new, optional "outputs" argument to the name_new,
name_firstupdate and name_update commands.  It follows the same semantics
as the "outputs" argument of the existing "paytomany" command.  With it,
name operations can be made that also include certain currency payments.

With this, we can achieve the same effect as with Namecoin Core's
"sendCoins" option that can optionally be passed to the same commands.