Allow zero outputs for auxpow coinbases.
When parsing the parent-block coinbase tx in an auxpow, accept it to
have no outputs at all.  While such transactions are invalid by
themselves, they are fine as part of an auxpow in Xaya Core (and may
be used if a "fake auxpow" is constructed when blocks are not
actually merge mined).