Fix delta computation for name transfers.
get_wallet_name_delta was broken for the situations of name transfers
(i.e. where one input is in the wallet and the other is not).  This was
due to a typo, where the domain check for name outputs was still done
on the (last) tx input address.

This change fixes the issue, and also makes the code a bit more readable
with extracting and sharing the long condition into a helper function.